socialproject/requirements.txt
argon2-cffi==21.1.0
arrow==1.2.0
asgiref==3.4.1
astroid==2.8.5
bcrypt==3.2.0
binaryornot==0.4.4
blis==0.7.5
botocore==1.21.65
catalogue==2.0.6
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
clean-text==0.5.0
click==8.0.3
cloudinary==1.26.0
colorama==0.4.3
convokit==2.5.1
cookiecutter==1.7.3
cryptography==35.0.0
cycler==0.11.0
cymem==2.0.6
defusedxml==0.7.1
dill==0.3.4
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.0.6
django-allauth==0.45.0
django-appconf==1.0.5
django-bootstrap3==15.0.0
django-braces==1.14.0
django-bulma==0.8.2
django-cors-headers==3.10.1
django-environ==0.7.0
django-quill-editor==0.1.22
django-sendgrid-v5==1.2.0
django-unused-media==0.2.2
djangorestframework==3.13.1
emoji==1.6.1
Faker==10.0.0
feedparser==6.0.8
fonttools==4.28.5
ftfy==6.0.3
future==0.16.0
gunicorn==20.1.0
idna==3.2
isort==5.10.1
Jinja2==3.0.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
langcodes==3.3.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
msgpack==1.0.3
msgpack-numpy==0.4.7.1
murmurhash==1.0.6
nltk==3.6.6
numpy==1.21.2
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
pathspec==0.5.9
pathy==0.6.1
Pillow==8.3.2
pip==21.2.4
platformdirs==2.4.0
poyo==0.5.0
preshed==3.0.6
psycopg2-binary==2.9.2
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
pylint==2.11.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.4
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.8.5
sendgrid==6.9.3
sgmllib3k==1.0.0
six==1.14.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.1
srsly==2.4.2
starkbank-ecdsa==2.0.3
termcolor==1.1.0
text-unidecode==1.3
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.0
Unidecode==1.3.2
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.1.9
wheel==0.37.0
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.13.3
